+++ cvrp_aco/vrpHelper.h
@@ -23,10 +23,10 @@
 #endif
 
 
-double compute_tour_length(Problem *instance, long int *t, long int t_sz);
-double compute_route_length(Problem *instance, long int *route, long int route_size);
+double compute_tour_length(Problem *instance, int *t, int t_sz);
+double compute_route_length(Problem *instance, int *route, int route_size);
 double **compute_distances(Problem *instance);
-long int ** compute_nn_lists (Problem *instance);
-void compute_route_centers(Problem *instance, long int *tour, const vector<RouteCenter *>& centers);
+int ** compute_nn_lists (Problem *instance);
+void compute_route_centers(Problem *instance, int *tour, const vector<RouteCenter *>& centers);
 
 #endif
